pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project.urls]
homepage = "https://github.com/archit-manek/football_pipeline"
repository = "https://github.com/archit-manek/football_pipeline"

[project]
name = "football_pipeline"
version = "0.1.0"
description = "Modular pipeline for football data and ML"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Archit Manek", email = "architmanek@gmail.com" }]
dependencies = [
  "pandas>=2.3",
  "pyarrow>=15",
  "numpy>=1.26",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "tqdm>=4.66",
  "requests>=2.32",
  "rich>=13.0",
  "joblib>=1.3",
  "socceraction[statsbomb]>=1.2",
  "polars>=1.30",
  "pandera>=0.17.2,<1.0.0",
  "multimethod<2.0"
]

[project.scripts]
football_pipeline = "football_pipeline.cli:main"

[project.optional-dependencies]
viz = ["matplotlib>=3.8", "seaborn>=0.13", "plotly>=5.20"]
ml = ["scikit-learn>=1.4"]
dev = ["pytest", "black", "ruff", "mypy"]

[tool.setuptools.packages.find]
where = ["src"]